Allow PHP 5.3 builds to fail
In order to support PHP 5.3, it would be necessary to downgrade to PHPCS 2.x, as 3.0-RC1 already bumps min PHP requirement to 5.4.
https://github.com/squizlabs/PHP_CodeSniffer/commit/161a3cfe6fb5c6cf324febd0f3c8395043cd35c2

However, PHPCS 2.x does not have support for PSR-12, which is the FIG standard, as it was only added in PHPCS 3.5
https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.5.0

This change should ensure that we can at least formally build on PHP 5.3, even if it cannot pass, and documents this problem in history.